DEP: Move _central_diff_weights & _derivative to _lib._finite_differences
